added grpc logger
