+++ hw12_13_14_15_calendar/configs/calendar.yml
@@ -15,4 +15,3 @@ dBItemsPerQuery: 50
 
 grpcAddress: "localhost:50051"
 httpAddress: "localhost:50052"
-httpProxyAddress: "localhost:50053"
