+++ hw12_13_14_15_calendar/go.mod
@@ -5,6 +5,7 @@ go 1.14
 require (
 	github.com/go-playground/validator/v10 v10.3.0
 	github.com/golang/protobuf v1.4.2
+	github.com/grpc-ecosystem/go-grpc-middleware v1.2.1
 	github.com/grpc-ecosystem/grpc-gateway v1.14.7
 	github.com/heetch/confita v0.9.2
 	github.com/imdario/mergo v0.3.11
@@ -14,7 +15,6 @@ require (
 	github.com/pressly/goose v2.6.0+incompatible
 	github.com/rs/zerolog v1.15.0
 	github.com/stretchr/testify v1.6.1
-	golang.org/x/net v0.0.0-20200226121028-0de0cce0169b // indirect
 	golang.org/x/text v0.3.3 // indirect
 	google.golang.org/genproto v0.0.0-20200815001618-f69a88009b70
 	google.golang.org/grpc v1.31.0
