+++ hw12_13_14_15_calendar/internal/app/app.go
@@ -2,6 +2,7 @@ package app
 
 import (
 	"errors"
+	"sync"
 
 	"github.com/dmitryt/otus-golang-hw/hw12_13_14_15_calendar/internal/config"
 	"github.com/dmitryt/otus-golang-hw/hw12_13_14_15_calendar/internal/repository"
@@ -19,18 +20,26 @@ func New(c *config.Config, r repository.Base) (*App, error) {
 	return &App{c: c, r: r}, nil
 }
 
-func (app *App) Run() error {
-	errCh := make(chan error)
+func (app *App) Run(errCh *chan error, doneCh *chan bool) {
 	s := service.New(app.r)
+
+	var wg sync.WaitGroup
+	wg.Add(1)
 	go func() {
+		defer wg.Done()
 		if app.c.GRPCAddress != "" {
-			errCh <- s.Run(app.c.GRPCAddress)
+			*errCh <- s.RunGRPC(app.c.GRPCAddress)
+		}
+	}()
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		if app.c.HTTPAddress != "" && app.c.GRPCAddress != "" {
+			*errCh <- s.RunHTTP(app.c.GRPCAddress, app.c.HTTPAddress)
 		}
 	}()
 	go func() {
-		if app.c.HTTPAddress != "" && app.c.GRPCAddress != "" {
-			errCh <- s.HTTPProxy(app.c.GRPCAddress, app.c.HTTPAddress)
-		}
+		wg.Wait()
+		close(*doneCh)
 	}()
-	return <-errCh
 }
